fix: implement EdgeUpgradeBanner as separate component
